requirements.txt
# Core
numpy>=1.21.0
pandas>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# Segmentation
segmentation-models-pytorch>=0.3.3

# Data
opencv-python>=4.5.0
rasterio>=1.2.0
scikit-learn>=0.24.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
matplotlib>=3.4.0
Pillow>=8.3.0

# Optional but good
jupyterlab>=3.0
tensorboard>=2.9.0
